Edit update - getting post data
getting post data in the form.
pushing the updatePost button attaches all to url- default issue
