+++ js/index.mjs
@@ -14,14 +14,13 @@ document.addEventListener("DOMContentLoaded", () => {
     listeners.setRegisterFormListener();
   } else if (path === "/profile/edit/") {
     listeners.setUpdateProfileListener();
-  } else if (path === `/feed/edit/`) {
-    listeners.setUpdatePostListener();
+ 
   } else if (path === "/feed/index.html") {
     listeners.displayPosts();
     listeners.setCreatePostListener();
-  }
+    listeners.setUpdatePostListener();
+}
 
-  // utils.profileUpdater.init();
-  // feedUpdater.init();
+ 
   utils.updateProfileFromLocalStorage();
 });
